requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
awscli==1.16.157
backcall==0.1.0
bleach==3.1.0
boto3==1.9.147
botocore==1.12.147
certifi==2019.3.9
cffi==1.12.3
Click==7.0
colorama==0.3.9
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
idna==2.8
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pyaml==19.4.1
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.0
PyMySQL==0.9.3
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==3.13
pyzmq==18.0.0
qtconsole==4.4.4
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.3
statsmodels==0.9.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wincertstore==0.2
